% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalvingPrep_main}
\alias{CalvingPrep_main}
\title{Main Entry Function To Check Data Consistency and Coding For Birth Evaluation}
\usage{
CalvingPrep_main(paramFileName)
}
\arguments{
\item{paramFileName}{input paramter file name}
}
\description{
This is the main entry function to check consistency and coding birth data.
}
